+++ addons/curved_lines_2d/line_2d_generator_inspector_plugin.gd
@@ -4,7 +4,7 @@ extends EditorInspectorPlugin
 class_name  Line2DGeneratorInspectorPlugin
 
 const GROUP_NAME_CURVE_SETTINGS := "Curve settings"
-
+const GROUP_NAME_EXPORT_OPTIONS := "Export Options"
 
 func _can_handle(obj) -> bool:
 	return obj is DrawablePath2D or obj is ScalableVectorShape2D
@@ -25,25 +25,6 @@ func _parse_begin(object: Object) -> void:
 		button.tooltip_text = "Pressing this button will change the way it is edited to Path mode."
 		add_custom_control(button)
 		button.pressed.connect(func(): _on_convert_to_path_button_pressed(object, button))
-	if object is ScalableVectorShape2D:
-		var button : Button = Button.new()
-		button.text = "Export as PNG*"
-		button.tooltip_text = "The export will only contain this node and its children,
-				assigned nodes outside this subtree will not be drawn."
-		add_custom_control(button)
-		button.pressed.connect(func(): _on_export_png_button_pressed(object))
-	if object is ScalableVectorShape2D:
-		var button : Button = Button.new()
-		button.text = "Export as baked scene*"
-		button.tooltip_text = "The export will only contain this node and its children,
-				assigned nodes outside this subtree will not be drawn.\n
-				⚠️ Warning:
-				- AnimationPlayer will not be included.
-				- Cutouts are not yet supported for this feature. Alternatively,
-				  you can manually unlock any generated polygons and lines
-				  and copy+paste them into a new scene."
-		add_custom_control(button)
-		button.pressed.connect(func(): _on_export_baked_scene_pressed(object))
 
 
 func _parse_group(object: Object, group: String) -> void:
@@ -51,7 +32,26 @@ func _parse_group(object: Object, group: String) -> void:
 		var key_frame_form = load("res://addons/curved_lines_2d/batch_insert_curve_point_key_frames_inspector_form.tscn").instantiate()
 		key_frame_form.scalable_vector_shape_2d = object
 		add_custom_control(key_frame_form)
+	elif group == GROUP_NAME_EXPORT_OPTIONS and object is ScalableVectorShape2D:
+		var box := VBoxContainer.new()
+		var export_png_button : Button = Button.new()
+		export_png_button.text = "Export as PNG*"
+		export_png_button.tooltip_text = "The export will only contain this node and its children,
+				assigned nodes outside this subtree will not be drawn."
+		export_png_button.pressed.connect(func(): _on_export_png_button_pressed(object))
+		var bake_button : Button = Button.new()
+		bake_button.text = "Export as baked scene*"
+		bake_button.tooltip_text = "The export will only contain this node and its children,
+				assigned nodes outside this subtree will not be drawn.\n
+				⚠️ Warning: An exported AnimationPlayer will not support animated curves"
+		box.add_theme_constant_override("separation", 5)
+		box.add_spacer(true)
+		box.add_child(export_png_button)
+		box.add_child(bake_button)
+		box.add_spacer(false)
 
+		add_custom_control(box)
+		bake_button.pressed.connect(func(): _on_export_baked_scene_pressed(object))
 
 func _parse_property(object: Object, type: Variant.Type, name: String, hint_type: PropertyHint, hint_string: String, usage_flags: int, wide: bool) -> bool:
 	if name == "line" and (object is  ScalableVectorShape2D):
@@ -76,6 +76,8 @@ func _parse_property(object: Object, type: Variant.Type, name: String, hint_type
 		var assign_nav_form = load("res://addons/curved_lines_2d/assign_navigation_region_inspector_form.tscn").instantiate()
 		assign_nav_form.scalable_vector_shape_2d = object as ScalableVectorShape2D
 		add_custom_control(assign_nav_form)
+	elif name == "show_export_options" and (object is ScalableVectorShape2D):
+		return true
 	return false
 
 
@@ -169,58 +171,46 @@ func _export_png(svs : ScalableVectorShape2D, filename : String, dialog : Node) 
 	sub_viewport.queue_free()
 
 
-func _export_baked_scene(svs : ScalableVectorShape2D, filepath : String, dialog : Node) -> void:
+static func _export_baked_scene(svs : ScalableVectorShape2D, filepath : String, dialog : Node) -> void:
 	dialog.queue_free()
-
-	# Let's temporarily modify the current branch so we can create the baked scene.
-	var svs_owner: Node = svs.owner
-	var svs_children: Array[Node] = svs.get_children()
-	var svs_ownership: Array[Node]
-	var replace_map: Dictionary[Node, Node]
-	var root := Node2D.new()
-	root.name = svs.name
-	replace_map[svs] = root
-
-	# Collect all nodes to be replaced.
-	while svs_children.size() > 0:
-		var child: Node = svs_children.pop_back()
-
-		if child is AnimationPlayer:
-			continue
-
-		svs_children.append_array(child.get_children())
-
-		# Store ownership so we can undo later.
-		if child.owner == svs_owner:
-			svs_ownership.append(child)
-
-		if child is ScalableVectorShape2D:
-			var node := Node2D.new()
-			node.name = child.name
-			node.unique_name_in_owner = child.unique_name_in_owner
-			node.transform = child.transform
-			replace_map[child] = node
-
-	# Do modifications and create scene.
-	for node in replace_map:
-		node.replace_by(replace_map[node], true)
-
-	for child in svs_ownership:
-		if child in replace_map:
-			replace_map[child].owner = root
-		else:
-			child.owner = root
-
+	var new_node := Node2D.new()
+	EditorInterface.get_edited_scene_root().add_child(new_node)
+	new_node.owner = EditorInterface.get_edited_scene_root()
+	var result := _copy_baked_node(svs, new_node, EditorInterface.get_edited_scene_root())
+	result.transform = Transform2D.IDENTITY
+	for node in result.get_children():
+		_recursive_set_owner(node, result, EditorInterface.get_edited_scene_root())
 	var scene := PackedScene.new()
-	scene.pack(root)
+	scene.pack(result)
 	ResourceSaver.save(scene, filepath, ResourceSaver.FLAG_NONE)
-
-	# Undo modifications and clear temporary nodes.
-	for node in replace_map:
-		replace_map[node].replace_by(node, true)
-		replace_map[node].queue_free()
-
-	for child in svs_ownership:
-		child.owner = svs_owner
-
+	new_node.queue_free()
 	EditorInterface.open_scene_from_path(filepath)
+
+
+static func _copy_baked_node(src_node : Node, dst_parent : Node, dst_owner : Node) -> Node:
+	if src_node is ScalableVectorShape2D and src_node.get_children().is_empty():
+		return null
+	var dst_node : Node = (
+		Node2D.new() if src_node is ScalableVectorShape2D else
+		ClassDB.instantiate(src_node.get_class())
+	)
+	dst_parent.add_child(dst_node)
+
+	for prop in src_node.get_property_list():
+		if src_node is ScalableVectorShape2D and prop.name == "script":
+			break
+		if prop.name in dst_node:
+			dst_node.set(prop.name, src_node.get(prop.name))
+
+	dst_node.owner = dst_owner
+	for ch in src_node.get_children():
+		_copy_baked_node(ch, dst_node, dst_owner)
+	return dst_node
+
+
+static func _recursive_set_owner(node : Node, new_owner : Node, root : Node):
+	if node.owner != root:
+		return
+	node.set_owner(new_owner)
+	for child in node.get_children():
+		_recursive_set_owner(child, new_owner, root)